+++ instrumentation/opentelemetry-instrumentation-boto/src/opentelemetry/instrumentation/boto/__init__.py
@@ -120,18 +120,10 @@ class BotoInstrumentor(BaseInstrumentor):
         with self._tracer.start_as_current_span(
             "{}.command".format(endpoint_name), kind=SpanKind.CONSUMER,
         ) as span:
+            span.set_attribute("endpoint", endpoint_name)
             if args:
                 http_method = args[0]
-                span.resource = Resource(
-                    attributes={
-                        "endpoint": endpoint_name,
-                        "http_method": http_method.lower(),
-                    }
-                )
-            else:
-                span.resource = Resource(
-                    attributes={"endpoint": endpoint_name}
-                )
+                span.set_attribute("http_method", http_method.lower())
 
             # Original func returns a boto.connection.HTTPResponse object
             result = original_func(*args, **kwargs)